- Throw a StravaException when an issue happen over the JSON conversion - Publish version 1.4

+++ src/main/java/org/jstrava/entities/LapsItem.java
@@ -35,7 +35,7 @@ public class LapsItem{
 	private boolean deviceWatts;
 
 	@SerializedName("total_elevation_gain")
-	private int totalElevationGain;
+	private double totalElevationGain;
 
 	@SerializedName("lap_index")
 	private int lapIndex;
@@ -143,7 +143,7 @@ public class LapsItem{
 		this.totalElevationGain = totalElevationGain;
 	}
 
-	public int getTotalElevationGain(){
+	public double getTotalElevationGain(){
 		return totalElevationGain;
 	}
